+++ components/popover/popover.vue
@@ -1,8 +1,12 @@
 <template>
-  <component :is="elementType">
+  <component
+    :is="elementType"
+    ref="popover"
+  >
     <dt-lazy-show
       v-if="modal"
-      :show="isOpeningPopover"
+      ref="overlay"
+      :show="modal && isOpeningPopover"
       transition="d-zoom"
       class="
         d-popover-overlay
@@ -82,11 +86,10 @@ import {
   POPOVER_PADDING_CLASSES,
   POPOVER_ROLES,
   POPOVER_VERTICAL_ALIGNMENT,
-  TIPPY_HORIZONTAL_ALIGNMENT,
 } from './popover_constants';
 import { getUniqueString } from '../utils';
 import DtLazyShow from '../lazy_show/lazy_show';
-import { TOOLTIP_HIDE_ON_CLICK_VARIANTS, TOOLTIP_TIPPY_DIRECTIONS } from '../tooltip';
+import { TOOLTIP_HIDE_ON_CLICK_VARIANTS } from '../tooltip';
 import ModalMixin from '../mixins/modal.js';
 
 export default {
@@ -328,6 +331,11 @@ export default {
       default: 300,
       validator: zIndex => !!Number(zIndex),
     },
+
+    overlayAppendTo: {
+      type: HTMLElement,
+      default: () => document.body,
+    },
   },
 
   emits: ['update:open'],
@@ -348,33 +356,30 @@ export default {
 
   computed: {
     fallbackPlacements () {
+      const verticalVariants = POPOVER_VERTICAL_ALIGNMENT.filter(alignment => alignment);
+      const horizontalVariants = POPOVER_HORIZONTAL_ALIGNMENT.filter(alignment => alignment);
       if (this.fixedAlignment === null && this.fixedVerticalAlignment === null) {
-        return POPOVER_VERTICAL_ALIGNMENT
-          .filter(verticalAlignment => verticalAlignment)
-          .map(verticalAlignment => TIPPY_HORIZONTAL_ALIGNMENT
-            .map(horizontalAlignment => `${verticalAlignment}-${horizontalAlignment}`))
-          .flat();
+        return verticalVariants.map(vertical =>
+          horizontalVariants.map(horizontal =>
+            this.getPlacement(vertical, horizontal)),
+        ).flat();
       }
 
       if (this.fixedAlignment === null) {
-        return TIPPY_HORIZONTAL_ALIGNMENT
-          .map(horizontalAlignment => `${this.fixedVerticalAlignment}-${horizontalAlignment}`);
+        return horizontalVariants
+          .map(horizontal => this.getPlacement(this.fixedVerticalAlignment, horizontal));
       }
 
       if (this.fixedVerticalAlignment === null) {
-        const horizontalAlignment = this.fixedAlignment === 'left' ? 'start' : 'end';
-        return POPOVER_VERTICAL_ALIGNMENT
-          .filter(verticalAlignment => verticalAlignment)
-          .map(verticalAlignment => `${verticalAlignment}-${horizontalAlignment}`);
+        return verticalVariants
+          .map(vertical => this.getPlacement(vertical, this.fixedAlignment));
       }
 
       return [];
     },
 
     placement () {
-      const verticalAlignment = this.fixedVerticalAlignment || 'bottom';
-      const horizontalAlignment = this.fixedAlignment === 'left' ? 'start' : 'end';
-      return `${verticalAlignment}-${horizontalAlignment}`;
+      return this.getPlacement(this.fixedVerticalAlignment, this.fixedAlignment);
     },
 
     isDialog () {
@@ -390,11 +395,9 @@ export default {
 
   watch: {
     placement (placement) {
-      if (TOOLTIP_TIPPY_DIRECTIONS[placement]) {
-        this.tip?.setProps({
-          placement,
-        });
-      }
+      this.tip?.setProps({
+        placement,
+      });
     },
 
     fallbackPlacements: {
@@ -432,7 +435,9 @@ export default {
     if (this.modal) {
       this.anchorEl.classList.add('d-zi-notification');
       zIndex = zIndex > 600 ? zIndex : 700;
+      this.appendOverlay();
     }
+
     // align popover content width when
     if (this.contentWidth === 'anchor') {
       window.addEventListener('resize', this.onResize);
@@ -440,7 +445,7 @@ export default {
     this.tip = tippy(this.anchorEl, this.getOptions({
       popperOptions: this.getPopperOptions(),
       tippyOptions: {
-        placement: this.getInitialPlacement(),
+        placement: this.placement,
         hideOnClick: this.hideOnClick,
         offset: this.offset,
         interactiveBorder: this.interactiveBorder,
@@ -462,13 +467,47 @@ export default {
 
   beforeDestroy () {
     window.removeEventListener('resize', this.onResize);
+    this.removeOverlay();
     this.tip?.destroy();
+    this.removeReferences();
   },
 
   /******************
    *     METHODS    *
    ******************/
   methods: {
+    removeReferences () {
+      this.anchorEl = null;
+      this.popoverContentEl = null;
+      this.tip = null;
+    },
+
+    appendOverlay () {
+      const overlay = this.$refs.overlay.$el;
+      const { lastChild } = this.overlayAppendTo;
+      if (lastChild) {
+        this.overlayAppendTo.insertBefore(overlay, lastChild);
+      } else {
+        this.overlayAppendTo.append(overlay);
+      }
+    },
+
+    removeOverlay () {
+      if (this.$refs.overlay && this.$refs.overlay.$el) {
+        this.$refs.popover.append(this.$refs.overlay.$el);
+      }
+    },
+
+    getPlacement (vertical = 'bottom', horizontal = 'end') {
+      const verticalAlignment = vertical || 'bottom';
+      const horizontalAlignment = horizontal || 'end';
+      if (horizontalAlignment === 'center') return verticalAlignment;
+      if (horizontalAlignment === 'left' || horizontalAlignment === 'right') {
+        return `${verticalAlignment}-${horizontalAlignment === 'left' ? 'start' : 'end'}`;
+      }
+      return `${verticalAlignment}-${horizontalAlignment}`;
+    },
+
     closePopover () {
       if (typeof this.hideOnClick === 'boolean' && this.hideOnClick) {
         this.tip.hide();
@@ -507,11 +546,9 @@ export default {
     onMount () {
       this.isPreventHidePopover = false;
       this.showPopover = true;
-      if (TOOLTIP_TIPPY_DIRECTIONS[this.placement]) {
-        this.tip?.setProps({
-          placement: this.placement,
-        });
-      }
+      this.tip?.setProps({
+        placement: this.placement,
+      });
     },
 
     onResize () {
@@ -538,9 +575,17 @@ export default {
               fallbackPlacements: this.fallbackPlacements,
             },
           },
+          {
+            name: 'hide',
+            enabled: this.appendTo !== 'parent',
+          },
           getArrowDetected(({ state }) => {
             this.verticalAlignment = state.placement.includes('top') ? 'top' : 'bottom';
-            this.horizontalAlignment = state.placement.includes('start') ? 'left' : 'right';
+            if (state.placement === 'top' || state.placement === 'bottom') {
+              this.horizontalAlignment = 'center';
+            } else {
+              this.horizontalAlignment = state.placement.includes('start') ? 'left' : 'right';
+            }
           }),
         ],
       };
@@ -563,10 +608,6 @@ export default {
       };
     },
 
-    getInitialPlacement () {
-      return TOOLTIP_TIPPY_DIRECTIONS[this.placement] ? this.placement : this.fallbackPlacements[0];
-    },
-
     async setPopoverContentAnchorWidth () {
       await this.$nextTick();
       this.popoverContentEl.style.width = `${this.anchorEl.clientWidth}px`;
@@ -586,10 +627,14 @@ export default {
 
 .dt-popover__content {
   &--align-right {
-    right: @su0;
-
     .dt-popover__caret {
       right: @su24;
+    }
+  }
+
+  &--align-center {
+    .dt-popover__caret {
+      left: 50%;
     }
   }