+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# [1.7.0](https://github.com/dialpad/dialtone-vue/compare/v1.6.1...v1.7.0) (2021-12-15)
+
+
+### Bug Fixes
+
+* popover component. remove overlay on before destroy ([#214](https://github.com/dialpad/dialtone-vue/issues/214)) ([3f88ad8](https://github.com/dialpad/dialtone-vue/commit/3f88ad8b8c062094073917ae71030120ed6ac096))
+
+
+### Features
+
+* add button group component ([#204](https://github.com/dialpad/dialtone-vue/issues/204)) ([99bbe17](https://github.com/dialpad/dialtone-vue/commit/99bbe1717627234d1dc7b091ec775403808daeed))
+
+## [1.6.1](https://github.com/dialpad/dialtone-vue/compare/v1.6.0...v1.6.1) (2021-12-03)
+
+
+### Bug Fixes
+
+* disable hide when append to is parent ([#207](https://github.com/dialpad/dialtone-vue/issues/207)) ([f478b3e](https://github.com/dialpad/dialtone-vue/commit/f478b3edc4ee4e306759b8e0d42ebbb4340aa733))
+* popover (props alignment between dt-popover and dt-tooltip) ([#205](https://github.com/dialpad/dialtone-vue/issues/205)) ([f935aad](https://github.com/dialpad/dialtone-vue/commit/f935aada1a5c23fc540c3479aab7cb605e47fcee))
+
 # [1.6.0](https://github.com/dialpad/dialtone-vue/compare/v1.5.0...v1.6.0) (2021-11-25)